requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
Flask==0.10.1
Flask-Markdown==0.3
flask-mongoengine==0.7.5
Flask-Script==2.0.5
Flask-WTF==0.14
gocardless-pro==0.2.2
-e git+https://github.com/richardjpope/govuk-jinja-components@929a494a555d4f6abc20d3c22d2edeb62b7ae020#egg=govuk_jinja_components
gunicorn==19.6.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
Markdown==2.6.6
MarkupSafe==0.23
mongoengine==0.10.6
pycrypto==2.6.1
pymongo==3.2.2
PySocks==1.7.1
pytz==2016.3
PyYAML==5.3
redis==2.10.5
requests==2.9.1
SecureMongoEngine==0.1.2
six==1.10.0
twilio==5.4.0
Werkzeug==0.11.5
WTForms==2.2.1
